requirements_dev.txt
-r requirements.txt

coverage
pytest~=6.2.4
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-django
pyyaml~=5.4.1
pycodestyle
pylint